Работает регулятор скорости

+++ src/node/encoder/topic_encoder.h
@@ -4,6 +4,6 @@
 
 struct EncoderTopic {
     TOPIC_FIELDS;
-    volatile int16_t value = 0;
-    volatile int16_t period = 0;
+    volatile int32_t value = 0;
+    volatile uint32_t period = 0;
 };
